Determine the number of pages upfront by the totalCount property

+++ src/main/scala/com/oldsneerjaw/TransactionPage.scala
@@ -5,11 +5,11 @@ import play.api.libs.json._
 /**
   * A single page of transactions.
   *
-  * @param totalCount the number of transactions contained within
+  * @param totalCount the total number of transactions from all pages
   * @param page The page number
   * @param transactions The collection of transactions
   */
-case class TransactionPage(totalCount: Long, page: Long, transactions: Seq[TransactionInfo])
+case class TransactionPage(totalCount: Int, page: Int, transactions: Seq[TransactionInfo])
 
 object TransactionPage {
   // Automatically converts a transaction summary page to/from JSON because the class' fields have the exact same names and types as in the
